+++ BridgeTrafficLoadSim/BridgeLane.h
@@ -15,18 +15,18 @@ public:
 	bool	operator<(const CBridgeLane& other);
 	
 	void	Update(double curTime);
-	void	setLaneNo(int LaneNo);
-	int		getLaneNo(void);
+	void	setLaneNo(size_t LaneNo);
+	size_t	getLaneNo(void);
 	void	addLoadEffect(CInfluenceLine IL, double weight);
 	void	AddVehicle(CVehicle* pVeh);
-	int		getIndex(void);
-	void	setIndex(int indx);
+	size_t	getIndex(void);
+	void	setIndex(size_t indx);
 	size_t	purgeVehicles(double curTime);
 	double	getLength(void);
 	void	setLength(double length);
 	double	getTimeNextVehOff(void) const;
 	double	setTimeNextVehOff(void);
-	double	getLoadEffect(int NoLE);
+	double	getLoadEffect(size_t NoLE);
 	double	getLeadVehPosition(void);
 	size_t	getNoVehs(void);
 	std::vector<CVehicle*> getVehicles(void);
@@ -38,13 +38,13 @@ private:
 	std::vector<CInfluenceLine> m_vInfLine;
 	std::vector<CVehicle*>		m_vVehicles;
 	std::vector<CAxle>			m_vAxles;
-	int		m_LaneNo;
+	size_t	m_LaneNo;
 	double	m_CurTime;
 	double	m_Length;
 	size_t	m_NoLE;
 	double	m_TimeNextVehOff;
 	//int		m_NoVehs;
-	int		m_Index;
+	size_t	m_Index;
 	double	m_LaneWidth;
 };
