+++ BridgeTrafficLoadSim/Event.h
@@ -20,17 +20,17 @@ class CEvent
 {
 public:
 	CEvent();
-	CEvent(int);
-	CEvent(int, size_t noEffects);
+	CEvent(size_t);
+	CEvent(size_t, size_t noEffects);
 	virtual ~CEvent();
 	
 	bool operator<(const CEvent& x);
 
-	CEffect&	getMaxEffect(int effNo);
-	CEffect&	getMinEffect(int effNo);
-	unsigned int	getNoVehicles();
-	unsigned int	getNoTrucks();
-	int			getID();
+	CEffect&	getMaxEffect(size_t effNo);
+	CEffect&	getMinEffect(size_t effNo);
+	size_t		getNoVehicles();
+	size_t		getNoTrucks();
+	size_t		getID();
 	double		getMaxTime();
 	double		getMaxEffectTime();
 	size_t		getNoEffects();
@@ -38,13 +38,13 @@ public:
 	std::string getTimeStr();
 	
 	void		setStartTime(double StartTime);	
-	void		setID(int id);
-	void		setMaxEffect(CEffect Eff, int i);
-	void		setMinEffect(CEffect Eff, int i);
+	void		setID(size_t id);
+	void		setMaxEffect(CEffect Eff, size_t i);
+	void		setMinEffect(CEffect Eff, size_t i);
 	void		setNoEffects(size_t noEffects);
-	void		setCurEffect(int ce);
+	void		setCurEffect(size_t ce);
 
-	void		writeEffect(int k, std::string file, bool trucks);
+	void		writeEffect(size_t k, std::string file, bool trucks);
 	void		writeToFile(std::string file);
 	void		AddSingleEffect(CEffect effect);
 	
@@ -55,17 +55,17 @@ private:
 	void setDefaults();
 	template <typename T> std::string to_string(T const& value);
 
-	int		m_CurEffect;
+	size_t	m_CurEffect;
 	size_t	m_NoEffects;
-	int		m_EventID;
+	size_t	m_EventID;
 	double	m_StartTime;
 
-	int		DAYS_PER_MT;
-	int		MTS_PER_YR;
-	int		HOURS_PER_DAY;
-	int		SECS_PER_HOUR;
-	int		MINS_PER_HOUR;
-	int		SECS_PER_MIN;
+	size_t	DAYS_PER_MT;
+	size_t	MTS_PER_YR;
+	size_t	HOURS_PER_DAY;
+	size_t	SECS_PER_HOUR;
+	size_t	MINS_PER_HOUR;
+	size_t	SECS_PER_MIN;
 };
 
 #endif // !defined(AFX_EVENT_H__0B74769F_32EB_4F04_BD76_F27A2F5FF946__INCLUDED_)
